+++ transactron/utils/fifo.py
@@ -1,7 +1,6 @@
 from amaranth import *
 from transactron import Method, def_method, Priority, TModule
-from transactron._utils import MethodLayout
-from transactron.utils._typing import ValueLike
+from transactron.utils._typing import ValueLike, MethodLayout
 
 
 class BasicFifo(Elaboratable):
