+++ transactron/utils/functions_pythonic.py
